pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fano"
version = "0.1.0"
description = "Multi-component platform for mathematical exploration and documentation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # HTTP clients
    "aiohttp>=3.9.0",
    "requests>=2.31.0",

    # Web frameworks
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "flask>=3.0.0",

    # Data validation
    "pydantic>=2.8.0",

    # LLM APIs (via OpenRouter - no provider-specific SDK required)
    # anthropic, openai SDKs optional - OpenRouter uses standard HTTP

    # Database
    "sqlalchemy>=2.0.0",

    # Configuration
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "rich>=13.0.0",
    "asyncio-throttle>=1.0.0",
    "markdown>=3.5.0",
    "python-dateutil>=2.8.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "pytest-benchmark>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "shared*",
    "llm*",
    "explorer*",
    "documenter*",
    "control*",
    "researcher*",
]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.html", "*.md"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["llm/tests", "tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
